Rename links.jsx to external-link.jsx.
