+++ src/app/Links.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import ExternalLink from '../components/links';
+import ExternalLink from '../components/external-link';
 import './styles/atoms.scss';
 
 const Links = () => (
